refactor: organize test by directory
