+++ tests/courier/courier-smoke.spec.ts
@@ -1,5 +1,5 @@
 import { expect, test, type Page } from '@playwright/test';
-import { LoginPage } from './pages/login.page';
+import { LoginPage } from '../pages/login.page';
 
 const acceptedOrderCode = 'pu4-qsc';
 const juanOrderCustomer = 'Juan Paredes';
