+++ tests/favorites/favorites.spec.ts
